+++ crates/chain-listener/Cargo.toml
@@ -21,6 +21,7 @@ eyre = { workspace = true }
 serde_json = { workspace = true }
 tokio = { workspace = true }
 server-config = { workspace = true }
+types = { workspace = true }
 
 fluence-libp2p = { workspace = true }
 tempfile = { workspace = true }
