+++ crates/fs-utils/Cargo.toml
@@ -12,3 +12,4 @@ tokio-stream = { workspace = true, features = ["fs"] }
 tracing = { workspace = true }
 thiserror = { workspace = true }
 futures-util = { workspace = true }
+types = { workspace = true }
