+++ crates/server-config/Cargo.toml
@@ -12,6 +12,7 @@ fluence-libp2p = { workspace = true, features = ["tokio"] }
 air-interpreter-fs = { workspace = true }
 peer-metrics = { workspace = true }
 fluence-keypair = { workspace = true }
+types = { workspace = true }
 log = "0.4.20"
 toml = "0.7.3"
