+++ crates/spell-event-bus/Cargo.toml
@@ -28,6 +28,7 @@ thiserror = { workspace = true }
 log = { workspace = true }
 fluence-spell-dtos = { workspace = true }
 peer-metrics = { workspace = true }
+types = { workspace = true }
 
 [dev-dependencies]
 libp2p = { workspace = true }
