+++ crates/types/Cargo.toml
@@ -7,3 +7,6 @@ edition = "2021"
 [dependencies]
 libp2p-identity = { workspace = true, features = ["peerid"] }
 serde = { workspace = true }
+
+[dev-dependencies]
+serde_json = { workspace = true }
