+++ particle-services/src/lib.rs
@@ -40,4 +40,4 @@ mod health;
 mod persistence;
 
 pub use app_services::ServiceInfo;
-pub use types::PeerScope;
+pub use types::peer_scope::PeerScope;
